Removed ENABLED_WUCLASS_X #defines.
Since we have the enabled_wuclasses.xml, it is no longer necessary to use #defines to turn wuclasses on or off in the vm. Now, everything gets compiled, but only the ones that are actually used in enabled_wuclasses.xml, and thus will be references from GENERATEDnative_wuclasses.c, will actually be included in the final image by the linker.