Refactored the code and added angularjs
Moved all html to static folder so it's easier to call template in
AngularJS.

Refactored the front end so that "get" returns a parsed html and
anything else is for manipulating states on the server.

Changed the layout of the application editor so one can see the list of
applications on the left of the screen while at the same time able to
deploy and change the FBP on the right.