Fix #116.
Changed the index from uint32_t to int32_t. Other option was to remove the index >=0 compare, but indexes really can be < 0. For example just indexing args[-1] in Helloworld.java as a test will compile just fine, and end up calling AALOAD with index -1 on the JVM stack.

(actually the old code actually worked just fine, since casting the -1 to a uint32 would just become a huge number, and it's highly unlikely we'll be using arrays with 4 billion elements, so we'd still get an array out of bounds exception)