Issue #133 Fix remaining issue in property pull. Before, it would only set the “needs pull” flag when a new object is created. But the objects for hardware are created before the application is loaded, so they never get the flag set, even if the application connected them to a remote source. Now it will also set the “needs pull” flag after loading the component map, for any existing objects.
However, it seems the master will generate initial value entries, even for properties that are linked to something else. Not sure if that should be legal or not. (maybe as a default value in case we can’t connect to the source?) When this happens, setting the value from the initial value table will remove the “needs pull” flag.