Issue #151
Renamed major/minor version in infusion header (which wasn't being used) to infusion format version and infusion version. The last can still be configured, but the format version is hard coded. The VM will check if the infusion format version matches the expected version, and panic otherwise.
The reason for this change is that on the rtc branch the infusion format has been changed. On Bolun's branch we didn't change the infusion itself, but changed the binary table format. All these changes make the deployed WuKong application incompatible with the VM on develop, so we need to remember to rebuild the infuser when switching between different branches. This way the VM will just panic with a clear message when we forget.