Fixed max stack depth count
There was a strange +1 in the infuser, and in the VM the offset for stack start was +2 for no reason.
Also added separate int and total stack count. We don't really need it, but it kept confusing me otherwise so better to have both (at the expense of only 1 extra byte in the header)