Removed unnecessary check on return type.
It doesn't matter if we return something different from what the method header promisses, as long as the stack is empty afterwards. If the return value is different from what the caller expects, it is either discarded or the caller receives garbage in the return value registers, which is incorrect but not unsafe.